Gallery component moar refactor
