+++ src/renderer/components/Gallery/index.ts
@@ -1,0 +1,3 @@
+import Gallery from './Gallery'
+
+export default Gallery
