+++ src/renderer/components/Gallery/types.ts
@@ -1,0 +1,10 @@
+import { GalleryAttachmentElementProps } from '../attachment/mediaAttachment'
+import { Type } from '../../backend-com'
+
+export type MediaTabKey = 'images' | 'video' | 'audio' | 'files' | 'webxdc_apps'
+
+export type GalleryElement = (
+  props: GalleryAttachmentElementProps & {
+    openFullscreenMedia: (message: Type.Message) => void
+  }
+) => JSX.Element
