+++ src/renderer/components/TabList/TabList.tsx
@@ -2,29 +2,35 @@ import React, { PropsWithChildren, useState, useEffect } from 'react'
 import classNames from 'classnames'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
 
-
 interface TabListProps {
   tabNames: string[]
   tabChangeCb: (newTab: string) => void
-}  
+}
 
-export default function TabList({ tabNames, tabChangeCb, children }: PropsWithChildren<TabListProps>) {
+export default function TabList({
+  tabNames,
+  tabChangeCb,
+  children,
+}: PropsWithChildren<TabListProps>) {
   const [currentTab, setCurrentTab] = useState<number>(0)
-  useEffect(() => tabChangeCb(tabNames[currentTab]), [currentTab])
-  const [tabs, extra] = (function() {
-      if (Array.isArray(children)) {
-        if (children.length === 2) {
-          return children
-        } else if (children.length == 1) {
-          return [children, null]
-          // will it really happen? --Farooq
-        } else {
-          throw new Error("TabList must have exactly 1 or 2 children")
-        }
-      } else {
+  useEffect(
+    () => tabChangeCb(tabNames[currentTab]),
+    [currentTab, tabChangeCb, tabNames]
+  )
+  const [tabs, extra] = (function () {
+    if (Array.isArray(children)) {
+      if (children.length === 2) {
+        return children
+      } else if (children.length == 1) {
         return [children, null]
+        // will it really happen? --Farooq
+      } else {
+        throw new Error('TabList must have exactly 1 or 2 children')
       }
-    })()
+    } else {
+      return [children, null]
+    }
+  })()
   const childToRender = Array.isArray(tabs) ? tabs[currentTab] : tabs
   return (
     <>
@@ -40,14 +46,18 @@ export default function TabList({ tabNames, tabChangeCb, children }: PropsWithCh
   )
 }
 
-
 interface TabListHeaderProps {
   currentTab: number
   onChangeTab: (newTab: number) => void
   tabs: string[]
 }
 
-function TabListHeader({ currentTab, onChangeTab, tabs, children }: PropsWithChildren<TabListHeaderProps>) {
+function TabListHeader({
+  currentTab,
+  onChangeTab,
+  tabs,
+  children,
+}: PropsWithChildren<TabListHeaderProps>) {
   return (
     <div className='tablist-header'>
       <ul>
