+++ src/renderer/components/TabList/index.ts
@@ -1,3 +1,3 @@
 import TabList from './TabList'
 
-export default TabList;
+export default TabList
